requirements.txt
Flask>=1.1.2
cx_Oracle==8.3.0
Jinja2>=2.11.3
python-dotenv==0.20.0
